♻️ refactor models, schemas, base class
