+++ backend/src/constants/role.py
@@ -1,0 +1,6 @@
+from sqlalchemy import Enum
+
+class UserRole(str, Enum):
+    admin = "admin"
+    user = "user"
+    moderator = "moderator"
