+++ backend/src/models/tag.py
@@ -1,0 +1,6 @@
+from src.models.base import Base
+from sqlalchemy import Column, String
+
+class Tag(Base):
+    __tablename__ = "tags"
+    name = Column(String, unique=True, index=True)
