+++ backend/src/models/user.py
@@ -1,0 +1,15 @@
+from sqlalchemy import Column, String, Boolean
+from src.models.base import Base
+
+
+class User(Base):
+
+    __tablename__ = 'users'
+    username = Column(String(50))
+    email = Column(String(250), nullable=False, unique=True)
+    password = Column(String(255), nullable=False)
+    avatar = Column(String(255), nullable=True)
+    refresh_token = Column(String(255), nullable=True)
+    confirmed = Column(Boolean, default=False)
+    active = Column(Boolean, default=False)
+    role = Column(String, default="user")
