+++ backend/src/schemas/email.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel, EmailStr
+
+class RequestEmail(BaseModel):
+    email: EmailStr
