+++ backend/src/schemas/tags.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+from typing import List
+
+
+class TagList(BaseModel):
+    tags: List[str]
