+++ backend/src/schemas/token.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class TokenModel(BaseModel):
+    access_token: str
+    refresh_token: str
+    token_type: str = "bearer"
