+++ backend/src/schemas/users.py
@@ -1,5 +1,4 @@
 from pydantic import BaseModel, EmailStr, Field
-from typing import List
 from datetime import datetime
 
 
@@ -24,17 +23,3 @@ class UserDb(BaseModel):
 class UserResponse(BaseModel):
     user: UserDb
     detail: str = "User successfully created"
-
-
-class TokenModel(BaseModel):
-    access_token: str
-    refresh_token: str
-    token_type: str = "bearer"
-
-
-class RequestEmail(BaseModel):
-    email: EmailStr
-
-
-class TagList(BaseModel):
-    tags: List[str]
